requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Click==8.1.7
Jinja2==3.1.2
psycopg2-binary==2.9.9
openai==1.43.0
openpyxl==3.1.5
numpy==1.24.3
pandas==2.0.3